+++ tests/image/test_uqi.py
@@ -170,5 +170,5 @@ def test_uqi_unequal_kernel_size():
         ]
     )
     # kernel order matters
-    assert universal_image_quality_index(preds, target, kernel_size=(3, 5)) == torch.tensor(0.10662283)
-    assert universal_image_quality_index(preds, target, kernel_size=(5, 3)) != torch.tensor(0.10662283)
+    torch.allclose(universal_image_quality_index(preds, target, kernel_size=(3, 5)), torch.tensor(0.10662283))
+    torch.allclose(universal_image_quality_index(preds, target, kernel_size=(5, 3)), torch.tensor(0.10662283))
